// design-mockup/src/lib/types/steel/Staafeinden.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Staafeind } from "./Staafeind";

/**
 * De twee staafeinden, in de referentierichting van de staaf (begin = x = 0).
 */
export type Staafeinden = { begin: Staafeind, eind: Staafeind, };